+++ packages/components/shared/src/events/select.ts
@@ -1,4 +1,5 @@
-export type SlSelectEvent<T = unknown> = CustomEvent<T>;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type SlSelectEvent<T = any> = CustomEvent<T>;
 
 declare global {
   interface GlobalEventHandlersEventMap {